refactor(agent): 拆分 embedding.py + 提取 resolve_model_info + 修复 6 个问题
1. models.py: 提取 resolve_model_info() 共享验证函数
   - build_chat_model 和 embedding/image_gen 统一使用
   - required_fields 参数化（chat 需要 base_url，image gen 不需要）

2. embedding.py 501→275 行（仅保留 embedding）
   - HybridEmbedding 改 frozen=True
   - _create_ark_client 去掉假 async
   - for_cluster = for_query（别名替代重复）
   - 删除 _resolve_model，改用 resolve_model_info

3. image_gen.py 217 行（新文件，图片生成）
   - 修复 _generate_image_openai httpx.AsyncClient 泄漏
   - 修复 base_url 语义不一致（Gemini 不强制要求）
   - 使用 resolve_model_info(model_id) 统一验证

4. 测试重组 + 补充
   - test_embedding.py 精简到 embedding 测试
   - test_image_gen.py 新增：dispatch + ark/openai/gemini helper 级测试
   - 覆盖 proxy 处理、reference_images、空响应错误路径
